+++ actionfx-core/src/main/java/com/github/actionfx/core/method/ControlValueAnnotatedParameterResolver.java
@@ -45,33 +45,46 @@ public class ControlValueAnnotatedParameterResolver implements AnnotatedParamete
 	@SuppressWarnings("unchecked")
 	@Override
 	public <T> T resolve(final Object controller, final Method method, final Parameter parameter,
-			final AFXControlValue controllValue, final Class<T> expectedType) {
+			final AFXControlValue controlValue, final Class<T> expectedType) {
 		final View view = ControllerWrapper.getViewFrom(controller);
 		if (view == null) {
 			throw new IllegalStateException("There is no view associated with controller of type '"
 					+ controller.getClass().getCanonicalName() + "'!");
 		}
-		final NodeWrapper control = view.lookupNode(controllValue.value());
+		final NodeWrapper control = view.lookupNode(controlValue.value());
 		if (control == null) {
-			throw new IllegalStateException("There is no node with ID='" + controllValue.value()
+			throw new IllegalStateException("There is no node with ID='" + controlValue.value()
 					+ "' inside the view associated with controller '" + controller.getClass().getCanonicalName()
 					+ "'!");
 		}
 		if (!javafx.scene.control.Control.class.isAssignableFrom(control.getWrappedType())) {
 			throw new IllegalStateException(
-					"Node with ID='" + controllValue.value() + "' inside the view hosted by controller '"
+					"Node with ID='" + controlValue.value() + "' inside the view hosted by controller '"
 							+ controller.getClass().getCanonicalName() + "' is not a javafx.scene.control.Control!");
 		}
 		final ControlWrapper controlWrapper = ControlWrapper.of(control.getWrapped());
-		final Object userValue = controlWrapper.getUserValue();
-		if (userValue != null && !parameter.getType().isAssignableFrom(userValue.getClass())) {
-			throw new IllegalStateException("User value retrieved for control with ID='" + controllValue.value()
-					+ "' inside the view hosted by controller '" + controller.getClass().getCanonicalName()
-					+ "' is not compatible with the method argument of type '" + parameter.getType().getCanonicalName()
-					+ "'! Control value is of type '" + userValue.getClass().getCanonicalName() + "'");
+		final Object value;
+		switch (controlValue.sourceProperty()) {
+		case SINGLE_VALUE_PROPERTY:
+			value = controlWrapper.getValue();
+			break;
+		case ITEMS_OBSERVABLE_LIST:
+			value = controlWrapper.getItems();
+			break;
+		case USER_VALUE_OBSERVABLE:
+			value = controlWrapper.getUserValue();
+			break;
+		default:
+			throw new IllegalStateException("Unsupported control property '" + controlValue.sourceProperty() + "'!");
+		}
+		if (value != null && !parameter.getType().isAssignableFrom(value.getClass())) {
+			throw new IllegalStateException("Value retrieved for control with ID='" + controlValue.value()
+					+ "' inside the view hosted by controller '" + controller.getClass()
+					+ "' is not compatible with the method argument of type '" + parameter.getType()
+					+ "'! Control value is of type '" + value.getClass() + "'");
 
 		}
-		return (T) userValue;
+		return (T) value;
 	}
 
 }